Implement antialiasing (#8)
* Opened fix_7 branch

* Open antialiasing branch

* Implement antialiasing

* Fix ImageTracer test

* FIx test
